obtain and display train data
this is the first version build that incorporates all the functionality that I want in the app. So areas of code need to be cleaned up, some UI/UX improvements need to be made, and more testing needs to be done in order to make sure there are no hidden bugs.